% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAS.R
\name{BAS}
\alias{BAS}
\title{Select points from a polygon using a BAS Master Sample.}
\usage{
BAS(
  shapefile,
  n = 100,
  boundingbox = NULL,
  panels = NULL,
  panel_overlap = NULL,
  stratum = NULL,
  nExtra = 1000,
  quiet = FALSE,
  seeds = NULL
)
}
\arguments{
\item{shapefile}{Shape file as a polygon (sp or sf) to select sites for.}

\item{n}{Number of sites to select. If using stratification it is a named vector containing
sample sizes of each group.}

\item{boundingbox}{Bounding box which defines the Master Sample. Default is BC Marine
Master Sample.}

\item{panels}{A list of integers that define the size of each panel in a
non-overlapping panels design. The length of the list determines the number of
panels required. The sum of the integers in the panels parameter will determine
the total number of samples selected, n. The default value for panels is NULL,
this indicates that a non-overlapping panel design is not wanted.}

\item{panel_overlap}{A list of integers that define the overlap into the previous
panel. Is only used when the panels parameter is not NULL. The default value for
panel_overlap is NULL. The length of panel_overlap must be equal to the length
of panels. The first value is always forced to zero as the first panel never
overlaps any region.}

\item{stratum}{Name of column of data.frame attached to shapefile that makes up the strata.}

\item{nExtra}{An efficiency problem of how many extra samples to draw before spatial clipping
to shp.}

\item{quiet}{Boolean if you want to see any output printed to screen. Helpful if taking a
long time.}

\item{seeds}{A list of seeds, u1 and u2, specified by the user.}
}
\value{
A master sample.
}
\description{
This is the main function for selecting sites using the BAS master
sample. It assumes that you have already defined the master sample using the
BoundingBox() function or will be selecting a marine master sample site in BC.
}
\details{
This function was first written by Paul van Dam-Bates for the
package BASMasterSample.
}
\examples{
\dontrun{
data(Fed_MPAs_clipped)
# Sample sizes for each stratum:
# chose how many sites in each polygon in the dataset
N_Zone <- c("Adaptive Management Zone" = 30, "Marine" = 20, "Other" = 40)
# Rename the NA value as the function does not accept NA at the moment.
Fed_MPAs_clipped$ZONEDESC_E[is.na(Fed_MPAs_clipped$ZONEDESC_E)] <- "Other"
# Core Protection is totally within Adaptive Management Zone. Remove it or
# make it explicit that it is different.
shp.MPAs <- Fed_MPAs_clipped[Fed_MPAs_clipped$ZONEDESC_E != "Core Protection Zone", ]
# Select the Master Sample sites:
smp.str <- uc511::BAS(shp.MPAs, N = N_Zone, stratum = "ZONEDESC_E", quiet = FALSE)
plot(sf::st_geometry(shp.MPAs))
plot(sf::st_geometry(smp.str), add = T, col= "red", pch = 16)
}

}
