% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASMastersample.R
\name{BASMastersample}
\alias{BASMastersample}
\title{BASMastersample.R
Do something useful.}
\usage{
BASMastersample(parm1)
}
\arguments{
\item{parm1}{A parameter.}
}
\value{
something useful.
}
\description{
A description of this useful function.
}
