% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{BuildMS}
\alias{BuildMS}
\alias{buildMS}
\title{Build a new Master Sample with a random rotation and seed.}
\usage{
buildMS(shp, d = 2, showOutput = TRUE, rotate = TRUE)
}
\arguments{
\item{shp}{Spatial feature that defines the boundary of the area to define a bounding box over.}

\item{d}{Dimension of the new Master Sample, at this stage we only work with d=2.}

\item{showOutput}{Print the rotation and random seed when it is generated.}

\item{rotate}{Boolean of whether or not to randomly rotate the bounding box.}
}
\value{
bounding box for a master sample.
}
\description{
Randomly generate a seed from 10,000 possible values in right now 2 dimensions.
Note that in van Dam-Bates et al. (2018) we required that the random seed falls into main object shape, such
as one of the islands in New Zealand, or within marine environment for BC west coast. However, with a random rotation,
we are able to ignore that detail. If this function is used without a random rotation, we recommend running it until
the first master sample point does indeed fall within the largest scale of the master sample use.
}
\details{
This function was first written by Paul van Dam-Bates for the
package BASMasterSample and later ported to this package, uc511.
}
\examples{
\dontrun{
data(NS_bioregion)
bb <- uc511::buildMS(shp = NS_bioregion)         # Vertically aligned master sample bounding box.
bb.rot <- uc511::rotate.shp(bb, bb, back = TRUE) # Actual bounding box.
plot(sf::st_geometry(NS_bioregion))
plot(sf::st_geometry(bb.rot), add = TRUE)
}

}
