% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compareBoxesBoxInit}
\alias{compareBoxesBoxInit}
\title{title.}
\usage{
compareBoxesBoxInit(boxes, boxInit, intB)
}
\arguments{
\item{boxes}{bla.}

\item{boxInit}{bla.}

\item{intB}{bla.}
}
\value{
A numeric vector.
}
\description{
description.
}
\details{
This function was written by Phil Davies.
}
