% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cppBASMasterSample}
\alias{cppBASMasterSample}
\title{A placeholder for now.}
\value{
Nothing at the moment.
}
\description{
A placeholder for now.
}
\details{
This function will be written by Phil Davies.
}
