% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cppHaltonSeq}
\alias{cppHaltonSeq}
\title{Draw Halton Sequence values for a single dimension.}
\usage{
cppHaltonSeq(k, base, n)
}
\arguments{
\item{k}{An integer for the starting index k >= 0.}

\item{base}{Co-prime Base but generally for BAS work it is 2 or 3.}

\item{n}{Number of samples to draw.}
}
\value{
something
}
\description{
Note that this was borrowed from the Internet and is not my implementation.
}
\examples{
cppHaltonSeq(k = 0, base = 2, n = 10)

}
