% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cppSumPoweredElements}
\alias{cppSumPoweredElements}
\title{Raise each element in a vector by a corresponding power provided in another vector,
then return the sum of all the results.}
\usage{
cppSumPoweredElements(J, bases, numElements)
}
\arguments{
\item{J}{A numeric vector of values with which to raise the corresponding
element in bases to.}

\item{bases}{A numeric vector containing values to raised to by the corresponding
powers in J}

\item{numElements}{The number of elements in the numeric vector bases.}
}
\value{
The sum of all the powers.
}
\description{
Raise each element in a vector by a corresponding power provided in another vector,
then return the sum of all the results.
}
