% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{getBB}
\alias{getBB}
\title{Get the bounding box for other functions.}
\usage{
getBB()
}
\value{
The bounding box.
}
\description{
Bounding box under NAD83/BC Albers.
}
\details{
This function was first written by Paul van Dam-Bates for the
package BASMasterSample.
}
