% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOverlappedPoints.R
\name{getOverlappedPoints}
\alias{getOverlappedPoints}
\title{Select points from a polygon using a BAS Master Sample.}
\usage{
getOverlappedPoints(shapefile, panelid)
}
\arguments{
\item{shapefile}{Shape file as a polygon (sp or sf) containing a sample that
contains a feature column named panel_id.}

\item{panelid}{The overlapped panel in the shapefile shp the user wants
sample points from.}
}
\value{
The sample for the specified panel.
}
\description{
This is the main function for selecting sites using the BAS master
sample. It assumes that you have already defined the master sample using the
BoundingBox() function or will be selecting a marine master sample site in BC.
}
\details{
This function was written by Phil Davies.
}
\examples{
\dontrun{
# Get all the sample from panel 1.
panelid <- 1
panel_1 <- uc511::getOverlappedPoints(shp, panelid)
}

}
