% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaltonFrames.R
\name{getSample}
\alias{getSample}
\title{Generate a Halton Frame.}
\usage{
getSample(shapefile, n, randomStart = FALSE)
}
\arguments{
\item{shapefile}{A MULTIPOINT or POINT object from where to take the sample.}

\item{n}{The number of sample points to return.}

\item{randomStart}{Whether a spatially balanced sample will be randomly drawn from
the frame or not. Default value is FALSE.}
}
\value{
A list containing the following variable:
sample - the sample from the shapefile POINTS.
}
\description{
A description of this useful function.
}
\details{
This function was written by Phil Davies.
}
\examples{
\dontrun{
hf_ <- uc511::getSample()
}

}
