% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{getSeed}
\alias{getSeed}
\title{Random seed definition for Western Canada Marine Master Sample.}
\usage{
getSeed()
}
\value{
The random seeds for the Western Canada Marine Master Sample.
}
\description{
Defines the random seed specific to the Western Canada Marine Master Sample.
}
\details{
This function was first written by Paul van Dam-Bates for the
package BASMasterSample and later ported to this package, uc511.
}
