% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaltonFrames.R
\name{HaltonFrames}
\alias{HaltonFrames}
\title{Do something useful.}
\usage{
HaltonFrames(parm1)
}
\arguments{
\item{parm1}{A parameter.}
}
\value{
something useful.
}
\description{
A description of this useful function.
}
