% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIP.R
\name{HIP}
\alias{HIP}
\title{A title.}
\usage{
HIP(population, n = 20, iterations = 7)
}
\arguments{
\item{population}{A population of points. (is the pop always pairs of points?).}

\item{n}{The number of points to draw from the population. Default 20.}

\item{iterations}{The levels of partitioning required. Default 7.}
}
\value{
A list containing the following variables:
sampleI - indices of the sample points to use.
popIndex - ???
Order - ???
HaltonIndex - Halton indices for all population points.
}
\description{
A description.
}
\details{
The details.
}
