% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIP.R
\name{is_sf_points}
\alias{is_sf_points}
\title{Check if an object is an sf points object.}
\usage{
is_sf_points(x)
}
\arguments{
\item{x}{A parm.}
}
\value{
Either TRUE or FALSE.
}
\description{
A description. internal only function.
}
\details{
The details.
}
