% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpacialBalanceMeasures.R
\name{SpacialBalanceMeasures}
\alias{SpacialBalanceMeasures}
\title{Do something useful.}
\usage{
SpacialBalanceMeasures(parm1)
}
\arguments{
\item{parm1}{A parameter.}
}
\value{
something useful.
}
\description{
A description of this useful function.
}
