% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialStratification.R
\name{SpatialStratification}
\alias{SpatialStratification}
\title{Do something useful.}
\usage{
SpatialStratification(parm1)
}
\arguments{
\item{parm1}{A parameter.}
}
\value{
something useful.
}
\description{
A description of this useful function.
}
