% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarianceEstimators.R
\name{VarianceEstimators}
\alias{VarianceEstimators}
\title{Do something useful.}
\usage{
VarianceEstimators(parm1)
}
\arguments{
\item{parm1}{A parameter.}
}
\value{
something useful.
}
\description{
A description of this useful function.
}
