% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAS.R
\name{BAS}
\alias{BAS}
\title{Draw spatially balanced samples from areal resources.}
\usage{
BAS(
  shapefile,
  n = 100,
  boundingbox = NULL,
  panels = NULL,
  panel_overlap = NULL,
  stratum = NULL,
  seeds = NULL,
  verbose = FALSE
)
}
\arguments{
\item{shapefile}{Shape file as a polygon (sp or sf) to select sites for.}

\item{n}{Number of sites to select. If using stratification it is a named vector containing
sample sizes of each group.}

\item{boundingbox}{Bounding box which defines the Master Sample. A bounding box must be
supplied.}

\item{panels}{A list of integers that define the size of each panel in a
non-overlapping panels design. The length of the list determines the number of
panels required. The sum of the integers in the panels parameter will determine
the total number of samples selected, n. The default value for panels is NULL,
this indicates that a non-overlapping panel design is not wanted.}

\item{panel_overlap}{A list of integers that define the overlap into the previous
panel. Is only used when the panels parameter is not NULL. The default value for
panel_overlap is NULL. The length of panel_overlap must be equal to the length
of panels. The first value is always forced to zero as the first panel never
overlaps any region.}

\item{stratum}{The name of a column in the data.frame attached to shapefile that defines
the strata of interest.}

\item{seeds}{A vector of 2 seeds, u1 and u2. If not specified, the default is NULL and will
be defined randomly using function \code{uc511::generateUVector}.}

\item{verbose}{Boolean if you want to see any output printed to screen. Helpful if taking a
long time. Default is FALSE i.e. no informational messages are displayed.}
}
\value{
A list containing two variables, \code{$sample} containing locations in the BAS sample,
in BAS order and \code{$seeds}, the u1 and u2 seeds used to generate the sample.

The sample points are returned in the form of a simple feature collection of POINT objects.
They have the following attributes:
\itemize{
\item \code{SiteID} A unique identifier for every sample point.  This
encodes the BAS order.
\item \code{geometry} The XY co-ordinates of the sample point in the CRS of the original
shapefile.
\item \code{uc511SeqID} A unique identifier for every sample point.  This
encodes the BAS sample order.
}
}
\description{
BAS draws spatially balanced samples from areal resources. To draw BAS samples,
uc511 requires a study region shapefile and the region’s bounding box. An initial sample size
is also needed, which can be easily increased or decreased within uc511 for master sampling
applications
}
\details{
This function was first written by Paul van Dam-Bates for the
package BASMasterSample and later simplified by Phil Davies.
}
\examples{
\dontrun{
data(Fed_MPAs_clipped)
# Sample sizes for each stratum:
# chose how many sites in each polygon in the dataset
N_Zone <- c("Adaptive Management Zone" = 30, "Marine" = 20, "Other" = 40)
# Rename the NA value as the function does not accept NA at the moment.
Fed_MPAs_clipped$ZONEDESC_E[is.na(Fed_MPAs_clipped$ZONEDESC_E)] <- "Other"
# Core Protection is totally within Adaptive Management Zone. Remove it or
# make it explicit that it is different.
shp.MPAs <- Fed_MPAs_clipped[Fed_MPAs_clipped$ZONEDESC_E != "Core Protection Zone", ]
# create a boundingbox
bb <- uc511::BoundingBox(shp.MPAs)
# Select the Master Sample sites:
smp.str <- uc511::BAS(shapefile = shp.MPAs,
                      n = N_Zone,
                      boundingbox = bb,
                      stratum = "ZONEDESC_E",
                      verbose = TRUE)
plot(sf::st_geometry(shp.MPAs))
plot(sf::st_geometry(smp.str$sample), add = T, col= "red", pch = 16)
plot(sf::st_geometry(bb), add = T)
}

}
