% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{getProj}
\alias{getProj}
\title{Define spatial objects in projection of the master sample.}
\usage{
getProj()
}
\value{
Spatial object projection.
}
\description{
Default projection of the master sample. Needed for consistency for the entire bounding box.
}
\details{
This function was first written by Paul van Dam-Bates for the
package BASMasterSample.
}
