% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{getRotation}
\alias{getRotation}
\title{Returns the radians of rotation of the bounding box.}
\usage{
getRotation()
}
\value{
something
}
\description{
Defines the random rotation specific to the Western Canada Marine Master Sample.
}
