% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaltonFrames.R
\name{HaltonFrame}
\alias{HaltonFrame}
\title{Generate a Halton Frame.}
\usage{
HaltonFrame(
  n = (bases[1]^J[1]) * (bases[2]^J[2]),
  J = c(3, 2),
  bases = c(2, 3),
  shapefile = NULL,
  panels = NULL,
  panel_overlap = NULL,
  randomStart = FALSE
)
}
\arguments{
\item{n}{The number of points in the frame to generate.}

\item{J}{The number of grid cells. A list of 2 values. The default value is c(3, 2), we could also use c(5, 3).}

\item{bases}{Co-prime base for the Halton Sequence. The default value is c(2, 3).}

\item{shapefile}{A sf object. If the shapefile parameter is NULL then function
uc511::HaltonFrameBase is called directly.}

\item{randomStart}{Whether a spatially balanced sample will be randomly drawn from
the frame or not. Default value is FALSE.}
}
\value{
A list containing the following variables:
- halton_seq = hf_$halton_seq
- halton_seq_div = hf_$halton_seq_div
- Z = hf_$Z
- halton_frame = hf_$halton_frame
- J = c(J\link{1}+i, J\link{2}+i)
- sample = diff_
- pts.shp = pts.shp
- bb = bb.new
}
\description{
A description of this useful function.
}
\details{
This function was written by Phil Davies.
}
\examples{
\dontrun{
hf_ <- uc511::HaltonFrame()
}

}
