% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaltonFrames.R
\name{HaltonFrame}
\alias{HaltonFrame}
\title{Generate a Halton Frame.}
\usage{
HaltonFrame(
  n = (bases[1]^J[1]) * (bases[2]^J[2]),
  J = c(3, 2),
  bases = c(2, 3),
  shapefile = NULL,
  crs = NULL
)
}
\arguments{
\item{n}{The number of points in the frame to generate.}

\item{J}{The number of grid cells. A list of 2 values. The default value is c(3, 2), we could also use c(5, 3).}

\item{bases}{Co-prime base for the Halton Sequence. The default value is c(2, 3).}

\item{shapefile}{something}

\item{crs}{The co-ordinate reference system.}
}
\value{
A list containing the following four variables:
halton_seq -
halton_seq_div -
Z -
halton_frame -
}
\description{
A description of this useful function.
}
\details{
This function was written by Phil Davies.
}
\examples{
\dontrun{
hf_ <- uc511::HaltonFrame()
}

}
