% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaltonFrames.R
\name{HaltonFrames}
\alias{HaltonFrames}
\title{A title.}
\usage{
HaltonFrames(parm1)
}
\arguments{
\item{parm1}{A parameter.}
}
\value{
Something very useful.
}
\description{
A description of this useful function.
}
\details{
This function will be written by Phil Davies.
}
